Move timer and helper funcs out of main for order purposes
